app/requirements.txt
streamlit
torch
pillow
numpy
opencv-python
